+++ src/components/invitation-creation-form.jsx
@@ -7,6 +7,7 @@ import { preventDefault } from '../utils';
 import { createFreefeedInvitation } from '../redux/action-creators';
 import SendTo from './send-to';
 import { Throbber } from './throbber';
+import { SignInLink } from './sign-in-link';
 
 export const INVITATION_LANGUAGE_OPTIONS = {
   RUSSIAN: 'ru',
@@ -34,7 +35,7 @@ class InvitationCreationForm extends Component {
       return (
         <div className="content">
           <div className="alert alert-danger" role="alert">
-            You must <Link to="/signin">sign in</Link> or <Link to="/signup">sign up</Link> before
+            You must <SignInLink>sign in</SignInLink> or <Link to="/signup">sign up</Link> before
             visiting this page.
           </div>
         </div>