+++ src/redux/reducers.js
@@ -1997,3 +1997,10 @@ export function appUpdated(state = { version: null, updated: false }, action) {
 }
 
 export { donationAccount, donationLoadingStatus } from './reducers/donation-status';
+
+export function sidebarOpened(state = false, action) {
+  if (action.type === ActionTypes.OPEN_SIDEBAR) {
+    return action.payload;
+  }
+  return state;
+}